requirements/local.txt
-r shared.txt

flake8
django-debug-toolbar
